Add files via upload
I want to present among all my data enthusiast connections. This is a dashboard I created using Tableau that analyzes hypothetical data for employee attrition in an organization, and the analysis is from the perspective of the HR department of that organization.

These objectives aim to help organizations understand attrition patterns, identify areas for improvement, and develop targeted strategies to reduce employee turnover.